backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.1
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.1
sentence-transformers>=2.7.0
numpy>=1.26.4
scikit-learn>=1.5.1
anthropic>=0.34.0
requests>=2.32.3
httpx>=0.27.0
firecrawl-py>=0.0.13
helix-py>=0.1.0
